requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
cloudscraper==1.2.71
undetected-chromedriver==3.5.5
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
requests-html==0.10.0
playwright==1.40.0